# bar.openapi.yml
openapi: 3.0.3
info:
  title: OpenAPI - Bar
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  version: 1.0.11
servers:
  - url: http://localhost:8080
  - url: https://le-bar.telecomnancy.net/api/v1
...
paths:
...
  /carousel/images:
    get:
      description: 'Get carousel images'
      operationId: getCarouselImages
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarouselImage'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      tags:
        - carousel
    post:
      description: 'Add a carousel image'
      operationId: addCarouselImage
      requestBody:
        description: Carousel image object
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CarouselImageUpload'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarouselImage'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '409':
          description: 'Item already exists'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - carousel
  /carousel/images/{image_id}:
    get:
      description: 'Get a carousel image'
      operationId: getCarouselImage
      parameters:
        - name: image_id
          in: path
          description: ID of the image
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: ''
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: 'Image not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      tags:
        - carousel
    delete:
      description: 'Delete a carousel image'
      operationId: markDeleteCarouselImage
      parameters:
        - name: image_id
          in: path
          description: ID of the image
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: 'Success'
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Image not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - carousel
  /carousel/texts:
    get:
      description: 'Get carousel texts'
      operationId: getCarouselTexts
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarouselText'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      tags:
        - carousel
    post:
      description: 'Add a carousel text'
      operationId: addCarouselText
      requestBody:
        description: Carousel text object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarouselTextCreate'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarouselText'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '409':
          description: 'Item already exists'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - carousel
  /carousel/texts/{text_id}:
    delete:
      description: 'Delete a carousel text'
      operationId: markDeleteCarouselText
      parameters:
        - name: text_id
          in: path
          description: ID of the text
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: 'Success'
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Text not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - carousel
...
components:
  schemas:
    UUID:
      type: string
      format: uuid
      example: 4be0643f-1d98-573b-97cd-ca98a65347dd
    ErrorCodes:
      type: string
      enum:
        - bad_request
        - forbidden
        - account_not_found
        - qr_invalid
        - internal_server_error
        - not_authenticated
        - image_not_found
        - text_not_found
        - category_not_found
        - item_not_found
        - refill_not_found
        - transaction_not_found
      x-enum-varnames:
        - ErrBadRequest
        - ErrForbidden
        - ErrAccountNotFound
        - ErrQRInvalid
        - ErrInternalServerError
        - ErrNotAuthenticated
        - ErrImageNotFound
        - ErrTextNotFound
        - ErrCategoryNotFound
        - ErrItemNotFound
        - ErrRefillNotFound
        - ErrTransactionNotFound
    Messages:
      type: string
      enum:
        - Bad request
        - Not connected
        - Missing fields %v
        - Account already exists
        - Account not found
        - Account is not allowed to use ressource
        - Account cannot use ressource at the time being
        - Invalid QR Code nonce
        - Internal server error
        - You are not authenticated
        - Image does not exists
        - Text does not exists
        - Category does not exists
        - Item does not exists
        - Refill does not exists
        - Transaction does not exists
      x-enum-varnames:
        - MsgBadRequest
        - MsgNotConnected
        - MsgMissingFields
        - MsgAccountAlreadyExists
        - MsgAccountNotFound
        - MsgAccountNotAllowed
        - MsgAccountNotAvailable
        - MsgQRInvalid
        - MsgInternalServerError
        - MsgNotAuthenticated
        - MsgImageNotFound
        - MsgTextNotFound
        - MsgCategoryNotFound
        - MsgItemNotFound
        - MsgRefillNotFound
        - MsgTransactionNotFound
    HTTPError:
      type: object
      properties:
        message:
          $ref: '#/components/schemas/Messages'
        error_code:
          $ref: '#/components/schemas/ErrorCodes'
      required:
        - message
        - error_code
...
    CarouselTextCreate:
      type: object
      properties:
        text:
          type: string
          description: Text to display
        color:
          type: string
          description: Color of the text
          example: "#FFFFFF"
      required:
        - text
    CarouselText:
      type: object
      properties:
        id: 
          $ref: '#/components/schemas/UUID'
        text:
          type: string
          description: Text to display
        color:
          type: string
          description: Color of the text
          example: "#FFFFFF"
        deleted_at:
          type: integer
          format: uint64
        deleted_by:
          $ref: '#/components/schemas/UUID'
      required:
        - id
        - text
        - color
    CarouselImageUpload:
      type: object
      properties:
        image:
          type: string
          description: Image to display
          format: binary
      required:
        - image
    CarouselImage:
      type: object
      properties:
        id: 
          $ref: '#/components/schemas/UUID'
        image_url:
          type: string
          description: Image to display
        deleted_at:
          type: integer
          format: uint64
        deleted_by:
          $ref: '#/components/schemas/UUID'
      required:
        - id
        - image_url
  securitySchemes:
    auth:
      type: apiKey
      name: BAR_SESS
      in: cookie
    local_token:
      type: apiKey
      name: X-Local-Token
      in: header
    admin_auth:
      type: apiKey
      name: BAR_ADMIN_SESS
      in: cookie